Move liberty settings copy from InstallServer to CreateServer.
